+++ app/components/Setting/actions.js
@@ -1,7 +1,7 @@
 import { createAction } from 'redux-actions'
 
 export const SETTING_ACTIVATE_TAB = 'SETTING_ACTIVATE_TAB'
-export const activateSettingsTab = createAction(SETTING_ACTIVATE_TAB)
+export const activateSettingTab = createAction(SETTING_ACTIVATE_TAB)
 
 export const SETTING_UPDATE_FIELD = 'SETTING_UPDATE_FIELD'
 export const updateSettingItem = createAction(SETTING_UPDATE_FIELD,
