+++ app/components/Setting/reducer.js
