+++ app/store.js
@@ -14,7 +14,7 @@ import TerminalReducer from './components/Terminal/reducer'
 import GitReducer from './components/Git/reducer'
 import WorkspaceReducer from './components/Workspace/reducer'
 import DragAndDropReducer from './components/DragAndDrop/reducer'
-import SettingReducer from './components/Settings/reducer'
+import SettingReducer from './components/Setting/reducer'
 
 const combinedReducers = combineReducers({
   FileTreeState: FileTreeReducer,
